# backend/openapi.yml
openapi: 3.0.0
info:
  title: Flask SQLAlchemy RESTful API
  version: 1.0.0
  description: |
    - 產生的 OpenAPI 文件，對應 Flask + SQLAlchemy 所有 RESTful API。
    
    - 所有 API 回傳 JSON。
servers:
  - url: http://127.0.0.1:8000
    description: Development server
...
paths:
  /user/{user_id}:
    get:
      tags: [User]
      summary: Get user by user_id
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  email:
                    type: string
                  created_at:
                    type: string
                    format: date
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "User not found"
  /register:
    post:
      tags: [User]
      summary: Register
      description: Register a new user with phone number, password, and email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone_number
                - password
                - email
              properties:
                phone_number:
                  type: string
                  pattern: '^[0-9]{8,15}$'
                  description: Phone number (8-15 digits only)
                  example: "0912345678"
                password:
                  type: string
                  minLength: 6
                  description: Password (minimum 6 characters)
                  example: "password123"
                email:
                  type: string
                  format: email
                  description: Email address (required)
                  example: "user@example.com"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Bad request - missing required fields or invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      missing_params:
                        value: "Phone number, password and email are required"
                      invalid_phone:
                        value: "Phone number must contain only digits"
                      phone_length:
                        value: "Phone number must be between 8 and 15 digits"
                      password_length:
                        value: "Password must be at least 6 characters long"
                      invalid_email:
                        value: "Invalid email format"
        '409':
          description: Conflict - phone number or email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      phone_exists:
                        value: "Phone number already exists"
                      email_exists:
                        value: "Email already exists"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Registration Failed"
  /login:
    post:
      tags: [User]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone_number
                - password
              properties:
                phone_number:
                  type: string
                  example: "0912345678"
                password:
                  type: string
                  example: "password123"
      responses:
        '200':
          description: User login successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Already logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyLoggedInResponse'
        '404':
          description: Missing required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Lack of necessary parameter"
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "phone number or password not correct"
...
  /current_user:
    get:
      tags: [User]
      summary: Get current logged in user
      description: Retrieve information about the currently logged in user from session
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Not logged in"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
  /user/profile:
    put:
      tags: [User]
      summary: Update user profile
      description: "Update current logged-in user's profile (email and password only). Note: If password is changed, user will be automatically logged out for security."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: New email address
                  example: "newemail@example.com"
                password:
                  type: string
                  description: New password (minimum 6 characters)
                  example: "newpassword123"
              description: "At least one field (email or password) must be provided"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      email_updated:
                        value: "Profile updated successfully"
                      password_updated:
                        value: "Password updated successfully. Please login again."
                  user:
                    $ref: '#/components/schemas/User'
                  logout:
                    type: boolean
                    description: True if user was logged out (when password changed)
                    example: true
        '400':
          description: Bad request (invalid data, no fields provided, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      no_data:
                        value: "No data provided"
                      no_fields:
                        value: "No valid fields to update (email or password)"
                      invalid_email:
                        value: "Invalid email format"
                      short_password:
                        value: "Password must be at least 6 characters long"
        '401':
          description: User not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Please login first"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email already exists"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to update profile"
  /update/username:
    put:
      tags: [User]
      summary: Update username
      description: Update the current logged-in user's username.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  description: New username
                  example: "User1234"
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Username updated successfully"
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request (invalid data, no username provided, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      no_data:
                        value: "No data provided"
                      invalid_username:
                        value: "Invalid username format"
        '401':
          description: User not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Please login first"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Username already exists"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to update username"
...
components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        email:
          type: string
          example: "user@example.com"
        created_at:
          type: string
          format: date
          example: "2025-09-20"
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login Successfully"
        User:
          $ref: '#/components/schemas/User'
    AlreadyLoggedInResponse:
      type: object
      properties:
        message:
          type: string
          example: "Already logged in"
        User:
          $ref: '#/components/schemas/User'
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: "Register Successfully"
        User:
          type: string
          description: User representation string
          example: "<User 1>"
...